Update to 36.0, lots of changes here. -Updated to use Windows 8.1 SDK, enabled gamepad support, fixed ANGLE WebGL leak, added build tools to git repo, improved memory usage

+++ testing/marionette/client/marionette/geckoinstance.py
@@ -6,6 +6,7 @@ from copy import deepcopy
 import errno
 import platform
 import os
+import sys
 import time
 
 from mozprofile import Profile
@@ -15,13 +16,19 @@ from mozrunner import Runner
 class GeckoInstance(object):
 
     required_prefs = {"marionette.defaultPrefs.enabled": True,
-                      "marionette.defaultPrefs.port": 2828,
                       "marionette.logging": True,
                       "startup.homepage_welcome_url": "about:blank",
                       "browser.shell.checkDefaultBrowser": False,
                       "browser.startup.page": 0,
                       "browser.sessionstore.resume_from_crash": False,
-                      "browser.warnOnQuit": False}
+                      "browser.warnOnQuit": False,
+                      "browser.displayedE10SPrompt": 5,
+                      "browser.displayedE10SPrompt.1": 5,
+                      "browser.displayedE10SPrompt.2": 5,
+                      "browser.displayedE10SPrompt.3": 5,
+                      "browser.displayedE10SPrompt.4": 5,
+                      "browser.tabs.remote.autostart.1": False,
+                      "browser.tabs.remote.autostart.2": False}
 
     def __init__(self, host, port, bin, profile, app_args=None, symbols_path=None,
                   gecko_log=None, prefs=None):
@@ -37,6 +44,7 @@ class GeckoInstance(object):
 
     def start(self):
         profile_args = {"preferences": deepcopy(self.required_prefs)}
+        profile_args["preferences"]["marionette.defaultPrefs.port"] = self.marionette_port
         if self.prefs:
             profile_args["preferences"].update(self.prefs)
         if not self.profile_path:
@@ -46,35 +54,44 @@ class GeckoInstance(object):
             profile_args["path_from"] = self.profile_path
             profile = Profile.clone(**profile_args)
 
-        if self.gecko_log is None:
-            self.gecko_log = 'gecko.log'
-        elif os.path.isdir(self.gecko_log):
-            fname = "gecko-%d.log" % time.time()
-            self.gecko_log = os.path.join(self.gecko_log, fname)
+        process_args = {
+            'processOutputLine': [NullOutput()],
+        }
 
-        self.gecko_log = os.path.realpath(self.gecko_log)
-        if os.access(self.gecko_log, os.F_OK):
-            if platform.system() is 'Windows':
-                # NOTE: windows has a weird filesystem where it happily 'closes'
-                # the file, but complains if you try to delete it. You get a
-                # 'file still in use' error. Sometimes you can wait a bit and
-                # a retry will succeed.
-                # If all retries fail, we'll just continue without removing
-                # the file. In this case, if we are restarting the instance,
-                # then the new logs just get appended to the old file.
-                tries = 0
-                while tries < 10:
-                    try:
-                        os.remove(self.gecko_log)
-                        break
-                    except WindowsError as e:
-                        if e.errno == errno.EACCES:
-                            tries += 1
-                            time.sleep(0.5)
-                        else:
-                            raise e
-            else:
-                os.remove(self.gecko_log)
+        if self.gecko_log == '-':
+            process_args['stream'] = sys.stdout
+        else:
+            if self.gecko_log is None:
+                self.gecko_log = 'gecko.log'
+            elif os.path.isdir(self.gecko_log):
+                fname = "gecko-%d.log" % time.time()
+                self.gecko_log = os.path.join(self.gecko_log, fname)
+
+            self.gecko_log = os.path.realpath(self.gecko_log)
+            if os.access(self.gecko_log, os.F_OK):
+                if platform.system() is 'Windows':
+                    # NOTE: windows has a weird filesystem where it happily 'closes'
+                    # the file, but complains if you try to delete it. You get a
+                    # 'file still in use' error. Sometimes you can wait a bit and
+                    # a retry will succeed.
+                    # If all retries fail, we'll just continue without removing
+                    # the file. In this case, if we are restarting the instance,
+                    # then the new logs just get appended to the old file.
+                    tries = 0
+                    while tries < 10:
+                        try:
+                            os.remove(self.gecko_log)
+                            break
+                        except WindowsError as e:
+                            if e.errno == errno.EACCES:
+                                tries += 1
+                                time.sleep(0.5)
+                            else:
+                                raise e
+                else:
+                    os.remove(self.gecko_log)
+
+            process_args['logfile'] = self.gecko_log
 
         env = os.environ.copy()
 
@@ -88,9 +105,7 @@ class GeckoInstance(object):
             cmdargs=['-no-remote', '-marionette'] + self.app_args,
             env=env,
             symbols_path=self.symbols_path,
-            process_args={
-                'processOutputLine': [NullOutput()],
-                'logfile': self.gecko_log})
+            process_args=process_args)
         self.runner.start()
 
     def close(self):
